Hackathon/ObjectDetectionAPI/requirements.txt
# Required packages for Object Detection WebSocket Server
numpy>=1.21.0
opencv-python>=4.8.0
mediapipe>=0.10.0
flask>=2.2.0
python-socketio>=5.7.0
eventlet>=0.33.0
Pillow>=9.5.0